Move MULTI-ZIG-SW into custom_devices
