+++ src/lib/customTypes.ts
@@ -1,0 +1,5 @@
+/* eslint-disable no-unused-vars */
+/* eslint-disable @typescript-eslint/no-namespace */
+export type Dictionary<T> = {
+    [key: string]: T;
+};
